Split HUD layout helper for lint
