+++ TiltArena/Game/Weapons/GravityWellState.swift
@@ -1,0 +1,8 @@
+import CoreGraphics
+import Foundation
+
+struct GravityWellState {
+    let center: CGPoint
+    let enemyIDs: Set<Int>
+    var timeRemaining: TimeInterval
+}
